+++ R/confint.psFit.R
@@ -14,9 +14,7 @@
 #'   \code{prof} containing the Wald and profile likelihood confidence intervals
 #'   respectively for the shape parameter of the fitted zeta distribution is
 #'   returned. In general these should be relatively close to each other.
-#'   **NOTE** These values are for the \pkg{VGAM} parameterisation of the Zeta
-#'   distribution which uses \eqn{s^\prime = s - 1}{s' = s - 1}. This means they
-#'   can be used without alteration in \code{\link[VGAM]{dzeta}}. If a
+#'   These values use the fitPS standard zeta shape parameter alpha, with shape > 1. VGAM uses a shifted parameter internally, equal to shape - 1. If a
 #'   zero-inflated zeta model is used (i.e. \code{object} comes from a call to
 #'   \code{\link{fitZIDist}}) then list of a confidence regions is returned with
 #'   and element for each value of \code{level}. The confidence regions are
@@ -69,13 +67,13 @@ confint.psFit = function(object, parm, level = 0.95, ...){
 
     profLik = Vectorize(
       function(bd){
-        2 * (sum(VGAM::dzeta(rep(obsData, object$psData$data$rn), shape = bd, log = TRUE)) +
+        2 * (sum(dzetaStandard(rep(obsData, object$psData$data$rn), shape = bd, log = TRUE)) +
            object$fit$value) + cstar
       })
 
     zstar = qnorm((1 - level) * 0.5, lower.tail = FALSE)
     wald = object$shape + c(-1,1) * zstar * sqrt(object$var.shape)
-    xmin = max(0, wald[1] - 0.5 * sqrt(object$var.shape))
+    xmin = max(1 + sqrt(.Machine$double.eps), wald[1] - 0.5 * sqrt(object$var.shape))
     xmax = wald[2] + 0.5 * sqrt(object$var.shape)
 
     lb = uniroot(profLik, c(xmin, object$shape), extendInt = "yes")$root