+++ R/print.psFit.R
@@ -16,9 +16,9 @@ print.psFit = function(x, ...){
 
   if(x$model %in% c("zeta", "ziz")){
     if(isBayes){
-      cat(paste("The estimated posterior mean of shape parameter is", round(x$shape + 1, 4), "\n"))
+      cat(paste("The estimated posterior mean of shape parameter is", round(x$shape, 4), "\n"))
     }else{
-      cat(paste("The estimated shape parameter is", round(x$shape + 1, 4), "\n"))
+      cat(paste("The estimated shape parameter is", round(x$shape, 4), "\n"))
     }
   }
 
@@ -36,7 +36,7 @@ print.psFit = function(x, ...){
 
   if(x$model %in% c("zeta", "ziz")){
     cat("------\n")
-    writeLines(strwrap("NOTE: The shape parameter is reported so that it is consistent with Coulson et al. However, the value returned is actually s' = shape - 1 to be consistent with the VGAM parameterisation, which is used for computation. This has flow on effects, for example in confInt. This will be changed at some point.\n"))
+    writeLines(strwrap("NOTE: The shape parameter is the standard zeta parameter alpha and must be greater than 1. VGAM uses a shifted parameter internally, equal to shape - 1, and fitPS converts to that value only at VGAM boundaries.\n"))
     cat("------\n\n")
 
     args = list(...)
@@ -47,7 +47,7 @@ print.psFit = function(x, ...){
         stop("nterms must be >= 1")
       }else if(nterms > 10){
         nvals = 1:nterms
-        fitted = VGAM::dzeta(nvals, shape = x$shape)
+        fitted = dzetaStandard(nvals, shape = x$shape)
         names(fitted) = if(x$type == 'P'){
           paste0("P", nvals - 1)
         }else{