+++ R/rZIzeta.R
@@ -4,7 +4,7 @@
 #' @param pi the mixing parameter for the zero-inflated zeta model---must be in
 #'   (0, 1).
 #' @param shape the shape parameter for the zero-inflated zeta. Must be greater
-#'   than zero.
+#'   than 1.
 #' @param offset the zeta distribution returns random variates that are greater
 #'   than, or equal to one. If the offset is greater than 0, then the
 #'   distribution is anchored on (has minimum value of) \code{1 - offset}.
@@ -25,7 +25,7 @@
 #' x = rZIzeta(n = sum(roux$data$rn), pi = fit.zi$pi, shape = fit.zi$shape)
 #' table(x)
 #' @export
-rZIzeta = function(n, pi = 0.5, shape = 1, offset = 0){
+rZIzeta = function(n, pi = 0.5, shape = 2, offset = 0){
 
   if(length(pi) > 1 || length(shape) > 1){
     stop("This function does not currently support vector valued inputs for Pi or shape.")
@@ -41,15 +41,13 @@ rZIzeta = function(n, pi = 0.5, shape = 1, offset = 0){
     stop("Pi must be in (0, 1)")
   }
 
-  if(shape <= 0){
-    stop("shape must be greater than zero.")
-  }
+  validateZetaShape(shape)
 
   p = runif(n)
   x = p
   x[p < pi] = 1
   n = sum(p >= pi)
-  x[p >= pi] = VGAM::rzeta(n, shape)
+  x[p >= pi] = rzetaStandard(n, shape)
   return(x - offset)
 }